Updated check to make sure API and DB are connected

+++ controllers/docid.controller.js
@@ -5,10 +5,10 @@ import _ from 'lodash';
 import { objectToCamelCase } from '../models/base.model.js';
 import userModel from '../models/user2.model.js'
 import agingDocModel from '../models/agingDocument.model.js';
+import { connectedToApi, connectedToDatabase } from '../models/healthCheck.model.js';
 
 
-
-export const checkDB = async (docID, email) => {
+export const checkDBForMatch = async (docID, email) => {
     const user = await userModel.findByEmail(email);
 
     const sql = 'SELECT agingDocument.* FROM agingDocument WHERE userID = ?';
@@ -49,22 +49,26 @@ result: JSON aging doc
 export const retrieveAgingDoc = async (req, res, next) => {
   try {
 
-    const docID = req.params.docID;
-    const email = req.params.email;
-    const agingDoc = await agingDocModel.findById(docID);
-    const response = await checkDB(docID, email);
-    const found = response["foundStatus"];
+    const connectedToApiResult = connectedToApi();
+    const connectedToDatabaseResult = await connectedToDatabase();
 
-    if (agingDoc == undefined) {
-        res.status(404).send("Error 404: Aging Document not found.")
-    }
-    else if (!found){
-        res.status(403).send("Error 403: Document does not belong to user.")
-    }
-    else {
-      return res.json(agingDoc);
-    }
-    
+    if (connectedToApiResult && connectedToDatabaseResult) {
+      const docID = req.params.docID;
+      const email = req.params.email;
+      const agingDoc = await agingDocModel.findById(docID);
+      const response = await checkDBForMatch(docID, email);
+      const found = response["foundStatus"];
+
+      if (agingDoc == undefined) {
+          res.status(404).send("Error 404: Aging Document not found.")
+      }
+      else if (!found){
+          res.status(403).send("Error 403: Document does not belong to user.")
+      }
+      else {
+        return res.json(agingDoc);
+      }
+  }
   } catch(err) {
     next(err);
   }
@@ -83,21 +87,26 @@ result: Error 403
 *******************************************************************************/
 export const removeAgingDoc = async (req, res, next) => {
   try {
-    const docID = req.params.docID;
-    const email = req.params.email;
-    const agingDoc = await agingDocModel.findById(docID);
-    const response = await checkDB(docID, email);
-    const found = response["foundStatus"];
+    const connectedToApiResult = connectedToApi();
+    const connectedToDatabaseResult = await connectedToDatabase();
 
-    if (agingDoc == undefined) {
-      res.status(404).send("Error 404: Aging Document not found.")
+    if (connectedToApiResult && connectedToDatabaseResult) {
+      const docID = req.params.docID;
+      const email = req.params.email;
+      const agingDoc = await agingDocModel.findById(docID);
+      const response = await checkDBForMatch(docID, email);
+      const found = response["foundStatus"];
+
+      if (agingDoc == undefined) {
+        res.status(404).send("Error 404: Aging Document not found.")
+      }
+      else if (!found){
+        res.status(403).send("Error 403: Document does not belong to user.")
+      }
+      else {
+        await agingDocModel.deleteById(docID);
+        res.status(200).end();
     }
-    else if (!found){
-      res.status(403).send("Error 403: Document does not belong to user.")
-    }
-    else {
-      await agingDocModel.deleteById(docID);
-      res.status(200).end();
   }
   } catch(err) {
     next(err);
@@ -117,37 +126,41 @@ result: {
 *******************************************************************************/
 export const points = async (req, res, next) => {
   try {
-    const docID = req.params.docID;
-    const email = req.params.email;
-    const response = await checkDB(docID, email);
-    const found = response["foundStatus"];
-    const user = response["user"];
-    const agingDoc = await agingDocModel.findById(docID);
+    const connectedToApiResult = connectedToApi();
+    const connectedToDatabaseResult = await connectedToDatabase();
 
-    const sqlImg = 'SELECT agedImage.uri FROM agingDocument, agingResult, agedImage WHERE agingDocument.userID = ? AND agingDocument.id = ? AND agingResult.agingDocument = agingDocument.id AND agingResult.id = agedImage.resultID';
-      const resultImg = await db.raw(sqlImg, [user.id, docID]).then((sqlResults) =>  {
-        return(objectToCamelCase(sqlResults[0]));
-    });
+    if (connectedToApiResult && connectedToDatabaseResult) {
+      const docID = req.params.docID;
+      const email = req.params.email;
+      const response = await checkDBForMatch(docID, email);
+      const found = response["foundStatus"];
+      const user = response["user"];
+      const agingDoc = await agingDocModel.findById(docID);
 
-    if (agingDoc == undefined) {
-      res.status(404).send("Error 404: Aging Document not found.")
-    }
-    else if (!found){
-      res.status(403).send("Error 403: Document does not belong to user.")
-    }
-    else if (Object.keys(resultImg).length == 0) {
-      res.status(307).end();
-    }
-    else {
-      const imgUrl = resultImg['0'].uri;
+      const sqlImg = 'SELECT agedImage.uri FROM agingDocument, agingResult, agedImage WHERE agingDocument.userID = ? AND agingDocument.id = ? AND agingResult.agingDocument = agingDocument.id AND agingResult.id = agedImage.resultID';
+        const resultImg = await db.raw(sqlImg, [user.id, docID]).then((sqlResults) =>  {
+          return(objectToCamelCase(sqlResults[0]));
+      });
 
-      res.json({
-        "uri": imgUrl,
-        "age": agingDoc.age
-      })
-      // res.send("<img src = "+imgUrl+">");
+      if (agingDoc == undefined) {
+        res.status(404).send("Error 404: Aging Document not found.")
+      }
+      else if (!found){
+        res.status(403).send("Error 403: Document does not belong to user.")
+      }
+      else if (Object.keys(resultImg).length == 0) {
+        res.status(307).end();
+      }
+      else {
+        const imgUrl = resultImg['0'].uri;
+
+        res.json({
+          "uri": imgUrl,
+          "age": agingDoc.age
+        })
+        // res.send("<img src = "+imgUrl+">");
+      }
     }
-    
   } 
   catch(err) {
     next(err);
@@ -173,22 +186,27 @@ This is the "Aging Request Representation" that this method needs
 *******************************************************************************/
 export const aging = async (req, res, next) => {
   try {
-    const docID = req.params.docID;
-    const email = req.params.email;
-    const response = await checkDB(docID, email);
-    const found = response["foundStatus"];
-    const agingDoc = await agingDocModel.findById(docID);
+    const connectedToApiResult = connectedToApi();
+    const connectedToDatabaseResult = await connectedToDatabase();
 
-    if (agingDoc == undefined) {
-        res.status(404).send("Error 404: Aging Document not found.")
-    }
-    else if (!found){
-        res.status(403).send("Error 403: Document does not belong to user.")
-    }
-    else {
-      console.log("Aging request successfuly sent to AprilAge")
-      console.log(req.body) // You send an Aging Request Representation in JSON (sequence types) to AprilAge
-      return res.status(202).end();
+    if (connectedToApiResult && connectedToDatabaseResult) {
+      const docID = req.params.docID;
+      const email = req.params.email;
+      const response = await checkDBForMatch(docID, email);
+      const found = response["foundStatus"];
+      const agingDoc = await agingDocModel.findById(docID);
+
+      if (agingDoc == undefined) {
+          res.status(404).send("Error 404: Aging Document not found.")
+      }
+      else if (!found){
+          res.status(403).send("Error 403: Document does not belong to user.")
+      }
+      else {
+        console.log("Aging request successfuly sent to AprilAge")
+        console.log(req.body) // You send an Aging Request Representation in JSON (sequence types) to AprilAge
+        return res.status(202).end();
+      }
     }
   } catch(err) {
     next(err);
@@ -206,20 +224,25 @@ result: {
 *******************************************************************************/
 export const status = async (req, res, next) => {
   try {
-    const docID = req.params.docID;
-    const email = req.params.email;
-    const response = await checkDB(docID, email);
-    const found = response["foundStatus"];
-    const agingDoc = await agingDocModel.findById(docID);
+    const connectedToApiResult = connectedToApi();
+    const connectedToDatabaseResult = await connectedToDatabase();
 
-    if (agingDoc == undefined) {
-        res.status(404).send("Error 404: Aging Document not found.")
-    }
-    else if (!found){
-        res.status(403).send("Error 403: Document does not belong to user.")
-    }
-    else {
-      return res.status(200).json({"status": agingDoc.status});
+    if (connectedToApiResult && connectedToDatabaseResult) {
+      const docID = req.params.docID;
+      const email = req.params.email;
+      const response = await checkDBForMatch(docID, email);
+      const found = response["foundStatus"];
+      const agingDoc = await agingDocModel.findById(docID);
+
+      if (agingDoc == undefined) {
+          res.status(404).send("Error 404: Aging Document not found.")
+      }
+      else if (!found){
+          res.status(403).send("Error 403: Document does not belong to user.")
+      }
+      else {
+        return res.status(200).json({"status": agingDoc.status});
+      }
     }
   } catch(err) {
     next(err);
